requirements.txt
groq
python-dotenv
pyperclip
